Created KafkaDelegationTokenTest and moved delegation token related common logic there
